feat(mobile): init navigation linking
