+++ apps/mobile/src/core/routes/routes.ts
@@ -1,0 +1,11 @@
+import { LinkingOptions } from '@react-navigation/native';
+import * as Linking from 'expo-linking';
+
+export const linking: LinkingOptions<ReactNavigation.RootParamList> = {
+  prefixes: [Linking.createURL('/')],
+  config: {
+    screens: {
+      Welcome: 'welcome',
+    },
+  },
+};
